Media Edit Dialog improvements

+++ src/app/manage/media/media-overview/media-overview.component.ts
@@ -7,6 +7,10 @@ import {AppService} from "../../../state/app.service";
 import {AppQueries} from "../../../state/app.queries";
 import {ScreenAssigningDialogComponent} from "./screen-assigning-dialog/screen-assigning-dialog/screen-assigning-dialog.component";
 import {WebsocketService} from "../../../core/services/websocket.service";
+import {
+  ConfirmationsPayload,
+  SimpleConfirmationDialogComponent
+} from "../../../shared/components/simple-confirmation-dialog/simple-confirmation-dialog.component";
 
 @Component({
   selector: 'app-media-overview',
@@ -41,7 +45,18 @@ export class MediaOverviewComponent implements OnInit {
   }
 
   onDelete(clipId: string) {
-    this.service.deleteClip(clipId);
+    const dialogRef = this._dialog.open(SimpleConfirmationDialogComponent, {
+      data: {
+        title: 'Are you sure you want to delete this clip?',
+      } as ConfirmationsPayload
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.service.deleteClip(clipId);
+      }
+    });
+
   }
 
   onEdit(item: Clip) {
